requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
python-dotenv==1.0.0
flask-paginate==2024.4.12
email-validator==2.1.0
gunicorn==22.0.0
